backend/foodgram/requirements.txt
requests==2.26.0
django==2.2.16
psycopg2-binary==2.8.6
python-dotenv
djangorestframework-simplejwt==4.7.1
django-colorfield
djoser
gunicorn==20.0.4
drf_multiple_serializer==0.2.3